% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitycheck_dates.R
\name{qualitycheck_dates}
\alias{qualitycheck_dates}
\title{qualitycheck_dates}
\usage{
qualitycheck_dates(Producto, Lote, Fecha_inicio, Protocolo)
}
\arguments{
\item{Producto}{A character of the name of the product. Ej: "Aspirin"}

\item{Lote}{A character with the ID of the Lote}

\item{Fecha_inicio}{A character date with the format in numer day-month-year; ej: "01-08-2019"}

\item{Protocolo}{A character date with the ID of protocol if this exist}
}
\value{
This function return a .csv document with the name of the Pruduct_Lote.csv in the 
directory of work, this document include all dates.
}
\description{
This function create a .csv document with 
the dates of a Lot of product since the date of departure date
}
\examples{
qualitycheck_dates("Aspririn", "Z0194", "03-03-2019")
qualitycheck_dates("Metformin", "0001", "08-01-2019")

}
