% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitycheck_lotes.R
\name{qualitycheck_lots}
\alias{qualitycheck_lots}
\title{qualititycheck_lots}
\usage{
qualitycheck_lots(Producto, Lista_lotes, Lista_fechas)
}
\arguments{
\item{Producto}{A character of the name of the product. Ej: "Aspirin"}

\item{Lista_lotes}{A character vector with the ID of the Lote}

\item{Lista_fechas}{A character vector with date in the format in numer day-month-year; ej: "01-08-2019"}
}
\value{
A .csv document with th dates of all lots
}
\description{
qualititycheck_lots
}
\examples{
b <- qualitycheck_lotes("Aspirina", c("1a","2b","3c","4c"), c("10-09-2019", "20-03-2019","30-03-2019", "30-11-2019"))

}
